backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# PDF Generation
reportlab==4.0.9
Pillow==10.2.0

# Excel Import
openpyxl==3.1.2
pandas==2.2.0

# CORS
python-dotenv==1.0.0

# AI/ML (Fase 2)
scikit-learn==1.4.0
numpy==1.26.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Utils
python-dateutil==2.8.2
pytz==2024.1
